test(gc): cover #7596's tenured-proportional nursery cap term (#7592)
Also splits the inline charCodeAt lowering into a sibling module to stay
under the 2000-line file cap.
